+++ PSAP/VIEW/INVVIEW/FrmSpecialWarehouseReceipt.cs
@@ -752,6 +752,13 @@ namespace PSAP.VIEW.BSVIEW
         {
             try
             {
+                if (gridViewSWRList.GetFocusedDataRow().RowState != DataRowState.Added)
+                {
+                    if (MessageHandler.ShowMessageBox_YesNo("确定要删除当前选中的明细记录吗？") != DialogResult.Yes)
+                    {
+                        return;
+                    }
+                }
                 gridViewSWRList.DeleteRow(gridViewSWRList.FocusedRowHandle);
             }
             catch (Exception ex)
