+++ PSAP/VIEW/INVVIEW/FrmSpecialWarehouseWarrant.cs
@@ -754,6 +754,13 @@ namespace PSAP.VIEW.BSVIEW
         {
             try
             {
+                if (gridViewSWWList.GetFocusedDataRow().RowState != DataRowState.Added)
+                {
+                    if (MessageHandler.ShowMessageBox_YesNo("确定要删除当前选中的明细记录吗？") != DialogResult.Yes)
+                    {
+                        return;
+                    }
+                }
                 gridViewSWWList.DeleteRow(gridViewSWWList.FocusedRowHandle);
             }
             catch (Exception ex)
