+++ PSAP/VIEW/PURVIEW/FrmOrder.cs
@@ -898,6 +898,13 @@ namespace PSAP.VIEW.BSVIEW
         {
             try
             {
+                if (gridViewOrderList.GetFocusedDataRow().RowState != DataRowState.Added)
+                {
+                    if (MessageHandler.ShowMessageBox_YesNo("确定要删除当前选中的明细记录吗？") != DialogResult.Yes)
+                    {
+                        return;
+                    }
+                }
                 gridViewOrderList.DeleteRow(gridViewOrderList.FocusedRowHandle);
             }
             catch (Exception ex)
