+++ PSAP/VIEW/PURVIEW/FrmPrReq.cs
@@ -839,6 +839,13 @@ namespace PSAP.VIEW.BSVIEW
         {
             try
             {
+                if (gridViewPrReqList.GetFocusedDataRow().RowState != DataRowState.Added)
+                {
+                    if (MessageHandler.ShowMessageBox_YesNo("确定要删除当前选中的明细记录吗？") != DialogResult.Yes)
+                    {
+                        return;
+                    }
+                }
                 gridViewPrReqList.DeleteRow(gridViewPrReqList.FocusedRowHandle);
             }
             catch (Exception ex)
